+++ src/ws_server/index.ts
@@ -1,54 +1,16 @@
-import WebSocket from 'ws';
-import { Message } from '../types';
-import { parseDataFromClient } from '../helpers';
-import { handleRegistration } from '../controllers/regController';
-import {
-  getUpdatedRoomInfo,
-  handleAddUserToRoom,
-  handleCreateRoom,
-} from '../controllers/roomController';
-import { handelCreateGame } from '../controllers/gameController';
+import { WebSocketServer } from 'ws';
+import { messageHandler } from './messageHandler';
 
-export const wsClients = new Set<WebSocket>();
+const WS_PORT = 3000;
 
-function broadcastToAllClients(data: string) {
-  for (const client of wsClients) {
-    if (client.readyState === WebSocket.OPEN) {
-      client.send(data);
-    }
-  }
-}
+export const webSocketServer = () => {
+  const wss = new WebSocketServer({
+    port: WS_PORT,
+  });
 
-const broadcastUpdatedRoomInfo = () => {
-  broadcastToAllClients(getUpdatedRoomInfo());
-};
+  wss.on('connection', (ws) => {
+    ws.on('message', messageHandler(ws));
+  });
 
-export const messageHandler = (ws: WebSocket) => (rawData: WebSocket.RawData) => {
-  try {
-    wsClients.add(ws);
-    const message: Message = JSON.parse(rawData.toString());
-    console.log(`Received command "${message?.type}"`);
-
-    switch (message?.type) {
-      case 'reg': {
-        const dataFromClient = parseDataFromClient(message);
-        handleRegistration(ws, dataFromClient);
-        broadcastUpdatedRoomInfo();
-        break;
-      }
-      case 'create_room': {
-        handleCreateRoom(ws);
-        broadcastUpdatedRoomInfo();
-        break;
-      }
-      case 'add_user_to_room': {
-        const dataFromClient = parseDataFromClient(message);
-        const roomData = handleAddUserToRoom(ws, dataFromClient?.indexRoom);
-        handelCreateGame(roomData);
-        broadcastUpdatedRoomInfo();
-      }
-    }
-  } catch {
-    console.log('Something went wrong');
-  }
+  console.log('Start websocket on port ', WS_PORT);
 };
