+++ src/ws_server/wsClients.ts
@@ -1,0 +1,3 @@
+import WebSocket from 'ws';
+
+export const wsClients = new Set<WebSocket>();
